refactor: 移除冗余类型标注并简化代码
本次提交主要进行了以下代码优化：
1. 移除整型和浮点型字面量的显式类型后缀（如_i32/_f64）
2. 简化位运算和数值计算表达式
3. 改进布尔参数的类型声明（使用bool替代i32）
4. 优化数组索引类型为usize
5. 清理不必要的类型转换和临时变量

+++ src/read_nmea_gga.rs
@@ -10,8 +10,8 @@ pub fn parse_f64(num_string: &str) -> Result<f64, std::num::ParseFloatError> {
 pub fn read_nmea_gga(
     xyz: &mut [[f64; 3]; USER_MOTION_SIZE],
     filename: &PathBuf,
-) -> anyhow::Result<i32> {
-    let mut numd: i32 = 0;
+) -> anyhow::Result<usize> {
+    let mut numd: usize = 0;
 
     let content = fs::read_to_string(filename)?;
 
@@ -58,7 +58,7 @@ pub fn read_nmea_gga(
         xyz[i][0] = pos[0];
         xyz[i][1] = pos[1];
         xyz[i][2] = pos[2];
-        numd = i as i32;
+        numd = i;
         // if i >= USER_MOTION_SIZE - 1 {
         //     break;
         // }
